pyproject.toml
[project]
name = "arag"
version = "0.1.0"
description = "A-RAG: Agentic RAG evaluation on financial data (arXiv:2602.03442)"
requires-python = ">=3.11,<3.14"
dependencies = [
    "anthropic>=0.45.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.0.0,<4.0.0",
    "transformers>=4.40.0,<5.0.0",
    "torch>=2.2.0,<3.0.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "streamlit>=1.35.0",
    "huggingface-hub>=0.24.0",
    "datasets>=2.20.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "ragas>=0.1.0",
    "sec-edgar-downloader>=5.0.0",
    "pandas>=2.2.0",
    "pytest>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
